Adjust Bikeshed level
The spec was published without level, meaning that it was published under
`picture-in-picture`, but not under `picture-in-picture-1`. This can be
adjusted later on if the group ever decides to work on a level 2, but no need
to do it now.

This update tells Bikeshed that the spec has no level, so that the publication
system knows that it has to publish under `picture-in-picture` and not under
`picture-in-picture-1`.